requirements.txt
numpy
setuptools
scipy
scikit-learn==0.20.3
opencv-python==4.2.0.32
opencv-contrib-python==4.2.0.32
tensorflow
keras_applications
tensorflow-model-optimization==0.5.0
tensorflow-addons
matplotlib
tqdm
pillow
mnn
Cython
pycocotools
keras2onnx
tf2onnx
onnx
onnxruntime
tfcoreml
sympy
imgaug
imagecorruptions
bokeh
tidecv